refactor(Db): Split the MySqlMapper in 3 Mappers
I splited the MySqlMapper in 3 mappers :

- FilesMapper : relative to the management files.
- MimeTypesMapper : relative to the MimeTypes database table.
- StoragesMapper : relative to the Storages database table.

I took the opportunity to sort lines ascending in constructors for each classes.
